db 바인딩

+++ app/src/main/java/com/example/snackbarcompose/data/models/Location.kt
@@ -12,7 +12,7 @@ data class Location(
     @ColumnInfo(name = "name")
     var name: String = "",
     @ColumnInfo(name = "lon")
-    var longitude: Float = 0.0F,
+    var longitude: Double = 0.0,
     @ColumnInfo(name = "lat")
-    var latitude: Float = 0.0F,
+    var latitude: Double = 0.0,
 )
